Create old_datapull_script.R
I'm waiting for IT to fix some Oracle issues that are preventing me from using Ben's package. In the meantime, I'm porting in the script I'd previously been using (inherited from Cole/Carey) that pulls the BSAI FHS data the old fashioned way.
Ultimate goal is to use afscdata for the pull but will need to do a comparison between the data anyway.