test run with old CAAL workup
this model uses survey CAALs which were generated by passing the ebs specimen data gathered from afscdata to Carey's code. I'm not a huge fan of the clunkiness of that syntax but the upshot is we were able to reproduce the input data faithfully.

Working up the afscdata pull from scratch resulted in wildly different comps (marginal and otherwise); a third thing I want to try is to investigate the gapindex package but we're currently working through some bugs.