test run with new survey marginal lengths
this model uses survey marginal lengths which were generated by passing the ebs length data gathered from afscdata to Carey's code. I'm not a huge fan of the clunkiness of that syntax and the downside is that I still wasn't able to get the data to look like the 2020 assessment, posing a serious issue. My guess is there must be some sort of filtration and/or expansion steps that I'm missing...but I really would like to get away from the bespoke code, and absolutely don't want to guess myself what was done before.

 I want to investigate the gapindex package but we're currently working through some bugs.