exploration of survey comp data sources
Did a closeup comparison of what came out of afscdata, the haehnr schema, and the new-and-improved gapindex package.
The high level takeaway is we CAN reproduce what was done in 2020 using the gapindex package, but not the afscdata. It is still unclear to me why  the $specimen data got used for the CAALs, and exactly how this differs from e.g. the $age or $length_at_age tables, and why POPULATION_COUNT isn't a factor there, but the point is we are able to retrace our steps.