finalize gapindex survey biomass method
thanks to ZO I was able to sort the area codes used to match the data. in the "deprecated" chunk at the bottom of this script is some ugly code used to compare the raw dwnld data for the ebs and ai; in the bridging analysis script I confirm that the results of the OM are in line with what was done before. big win!