EDA for otolith subsampling
A&G getting an early start on the reads and requested a spreadsheet of exactly which otoliths to read to get to the 375 per survey year. I was reassured by this look that randomly sampling (no weights, without replacement) the collections for each year would be sufficient. (The oto's are collected randomly by haul). I also wanted to avoid changing the data-generation process any more dramatically for this update year; in future years we could potentially explore things like forcing a number of reads by haul, etc.