more notes on the projection setup
the write_proj() function seems to duplicate the recruitment information and will cause an index error for spm. until I have time to correct that function make a note to remind reader to manually compare the input projection-data.dat files for errors. The versions from Apr 03 and May 29 2024 both work.