routine for interpolating NA values
even after updating to 500 samples there are still some year-bin combos with an NA sample size, generally on the fringes. per Pete's suggestion I look to the same sex-bin combo in the surrounding 10 years OR drop the sex requirement if that returns an NA as well. This allows for a dataframe that executes in SS.